[IMP] stock_available_to_promise_release: Own field for release policy
Currently if sale_stock_available_to_promise_release is installed
we are not able to change the release policy anymore.
Therefor we need a separate field,
which will be set from the procurement.group's move_type.

For Example we can now keep the shipping_policy on the transfer
but release only partially.