[FIX] st_available_to_promise_release: Fix location dest on cancel picking
When creating the return picking for moved goods on unrelease operations,
we are forcing 'Stock' (ie the default location of return picking type)
as a destination location, through context of return picking wizard.

However, this context key was only used on the preparation of the return
move values, whereas the return picking had its destination location set
to the source location of the picking being returned. Hence, when
the partner was set on the return picking (through purchase_stock override),
it was triggering recomputation of the destination location on the stock
move, overwriting the forced destination location from the return picking
type.

By using the location value provided through context to create the return
wizard when preparing the return picking values, we ensure any recomputation
will not overwrite said value on the stock move.