stock_available_to_promise_release: imp unrelease to backorder
It is possible to unrelease a single move from a transfer.
After doing so the transfer (or that single move) can be released again.

Before this change, when releasing the 2nd time, a bakorder is created
for the move not being release the 2nd time (because they are already released).

After this change, all released moves stay in the same transfer only the
non released move are moved in a backorder.